requirements.txt
requests==2.23.0
PySocks==1.7.1
PyYAML==5.3
redis==3.4.1
